AP-3493: Refactor in_scope_of_laspo controller
Move the controller, views and locale entries
remove mini loop testing as no longer needed
