+++ app/views/providers/application_merits_task/in_scope_of_laspos/show.html.erb
