+++ spec/requests/providers/application_merits_task/in_scope_of_laspos_controller_spec.rb
@@ -1,37 +1,36 @@
 require "rails_helper"
 
-RSpec.describe Providers::InScopeOfLasposController do
-  let(:legal_aid_application) { create(:legal_aid_application, :with_proceedings) }
+RSpec.describe Providers::ApplicationMeritsTask::InScopeOfLasposController do
+  let(:legal_aid_application) { create(:legal_aid_application, :with_multiple_proceedings_inc_section8) }
   let(:provider) { legal_aid_application.provider }
   let(:next_flow_step) { flow_forward_path }
+  let(:smtl) { create(:legal_framework_merits_task_list, :da001_as_defendant_and_child_section_8, legal_aid_application:) }
 
   before { login_as provider }
 
   describe "GET /providers/:application_id/in_scope_of_laspo" do
-    subject { get providers_legal_aid_application_in_scope_of_laspo_path(legal_aid_application) }
+    subject(:get_laspo) { get providers_legal_aid_application_in_scope_of_laspo_path(legal_aid_application) }
 
-    before { subject }
+    before { get_laspo }
 
     it "returns http success" do
       expect(response).to have_http_status(:ok)
     end
 
     it "displays the page" do
-      expect(unescaped_response_body).to include(I18n.t("providers.in_scope_of_laspos.show.page_title"))
+      expect(unescaped_response_body).to include(I18n.t("providers.application_merits_task.in_scope_of_laspos.show.page_title"))
     end
   end
 
   describe "PATCH /providers/:application_id/in_scope_of_laspo" do
-    subject { patch providers_legal_aid_application_in_scope_of_laspo_path(legal_aid_application), params: }
+    subject(:post_laspo) { patch providers_legal_aid_application_in_scope_of_laspo_path(legal_aid_application), params: }
 
     before do
-      allow(Setting).to receive(:enable_mini_loop?).and_return(mini_loop?)
-      subject
+      allow(LegalFramework::MeritsTasksService).to receive(:call).with(legal_aid_application).and_return(smtl)
+      post_laspo
     end
 
-    let(:mini_loop?) { false }
-
-    context "choose yes" do
+    context "when the user chooses yes" do
       let(:params) { { legal_aid_application: { in_scope_of_laspo: true } } }
 
       it "updates the record" do
@@ -43,7 +42,7 @@ RSpec.describe Providers::InScopeOfLasposController do
       end
     end
 
-    context "choose no" do
+    context "when the user chooses no" do
       let(:params) { { legal_aid_application: { in_scope_of_laspo: false } } }
 
       it "updates the record" do
@@ -53,18 +52,9 @@ RSpec.describe Providers::InScopeOfLasposController do
       it "redirects to the next page" do
         expect(response).to redirect_to(next_flow_step)
       end
-
-      context "when the mini_loop flag is on" do
-        let(:mini_loop?) { true }
-
-        it "redirects to the next page" do
-          proceeding_id = legal_aid_application.proceedings.first.id
-          expect(response).to redirect_to(providers_legal_aid_application_client_involvement_type_path(legal_aid_application.id, proceeding_id))
-        end
-      end
     end
 
-    context "choose nothing" do
+    context "when the user chooses nothing" do
       let(:params) { { legal_aid_application: { in_scope_of_laspo: nil } } }
 
       it "stays on the page if there is a validation error" do
@@ -73,7 +63,7 @@ RSpec.describe Providers::InScopeOfLasposController do
       end
     end
 
-    context "Form submitted with Save as draft button" do
+    context "when the form submitted with Save as draft button" do
       let(:params) { { legal_aid_application: { in_scope_of_laspo: false }, draft_button: "Save and come back later" } }
 
       it "redirects to the list of applications" do
